requirements.txt
python-dotenv
langchain
langchain_openai
langchain-community
langchain_text_splitters
langchain-chroma
langchain_experimental
unstructured[all-docs]
rank_bm25
langchain_cohere
scikit-learn
whoosh
